+++ 00.InfraStructure/00.D.PersistenceTest/PersistenceTest/Positions/PersistencePositionTestBuilder.cs
@@ -21,7 +21,7 @@ namespace PersistenceTest.Positions
     public string CustomesCode = someCustomesCode;
     public string Description = someDescription;
     public bool IsActive = someIsActive;
-    public int RoleId = someRoleId;
+    public long RoleId = someRoleId;
 
     public PersistencePositionTestBuilder()
     {
